int --> size_t  log2_boundary, random_depth

+++ sort.h
@@ -44,9 +44,9 @@ typedef int (*__compar_d_fn_t) (const void *, const void *, void *);
 /***** variables *****/
 
 extern size_t	INS;				// main.c
-extern int		random_depth;		// main.c
+extern size_t	random_depth;		// main.c
 extern double	random_number;		// main.c
-extern int		log2_boundary;		// main.c
+extern size_t	log2_boundary;		// main.c
 extern int		pivot_number;		// main.c
 extern bool		ispointer;			// index_sort.c
 extern double	usec, esd;			// timer.c
